Clarify intention of query result data handling
Intermediate value of result_data variable is a list of lists and not a list of RowDict objects. The QueryResult object will combine the headers and values into a list of RowDict objects - this allows for QueryResult objects to still to be created from arbitrary lists of headers and values.