Improve QueryGenerator and QueryExecutor
Extract result fetching into a staticmethod of QueryGenerator class, and only fetch when there are rows available. This allows QueryGenerator to be used for the execution of SQL which does not return rows (e.g. update, truncate, etc).

Modify QueryExecutor so that each instance is not restricted to just one query - within context of a given connection, can execute an arbitrary number of statements.

Add commit helper method to ConnectionManager and QueryExecutor.